Fix: add tests for all utils functions
