+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix: refactor parse issue header (@lwasser, #91)
 - Fix: Refactor and organize modules (@lwasser, #76)
 - Fix: Rename and organize `clean.py` module into `utils_parse` and `utils_clean` (@lwasser, @willingc, #121)
+- Fix: Add tests for all utils functions (@lwasser, #122)
 
 
 ## [v0.2.3] - 2024-02-29
